+++ packages/nimble-components/src/table/tests/table.stories.ts
@@ -133,7 +133,8 @@ const metadata: Meta<TableArgs> = {
             handles: [
                 'action-menu-beforetoggle',
                 'action-menu-toggle',
-                'selection-change'
+                'selection-change',
+                'column-configuration-change'
             ]
         }
     },